Extracted Header from Mainpage
