+++ frontend/src/containers/MainPage/MainPage.scss
@@ -23,55 +23,55 @@ $top: 20px;
     justify-content: space-around;
     flex-direction: row;
 }
-.MainPage .Header {
-    padding: 10px;
-    height: 125px;
-    filter: drop-shadow(0 10px 10px rgba(0, 0, 0, 0.1));
-    display: flex;
-    justify-content: space-evenly;
-    flex-direction: row;
-}
-.MainPage #logo-container {
-    width: auto;
-    display: flex;
-}
-.MainPage #logo-container {
-    align-items: center;
-}
-.MainPage #search-container {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 0px;
-}
-.MainPage #search-box-container {
-    flex-direction: row;
-    flex-wrap: nowrap;
-}
-.MainPage #search-box-container .search-box {
-    display: flex;
-    align-items: center;
-    margin-top: 5px;
-    font-family: "NanumGothic", serif;
-    font-size: 10px;
-    min-width: 800px;
-}
-.MainPage #button-container {
-    margin: 0;
-    text-align: left;
-    width: auto;
-    display: flex;
-    align-items: start;
-    justify-content: space-around;
-    margin-top: 15px;
-    min-width: 200px;
-}
-.MainPage #button-container .button {
-    font-size: $button-size;
-    // height: $button-size;
-    width: $button-size;
-    cursor: pointer;
-    z-index: 100;
-}
+// .MainPage .Header {
+//     padding: 10px;
+//     height: 125px;
+//     filter: drop-shadow(0 10px 10px rgba(0, 0, 0, 0.1));
+//     display: flex;
+//     justify-content: space-evenly;
+//     flex-direction: row;
+// }
+// .MainPage #logo-container {
+//     width: auto;
+//     display: flex;
+// }
+// .MainPage #logo-container {
+//     align-items: center;
+// }
+// .MainPage #search-container {
+//     display: flex;
+//     flex-direction: column;
+//     margin-bottom: 0px;
+// }
+// .MainPage #search-box-container {
+//     flex-direction: row;
+//     flex-wrap: nowrap;
+// }
+// .MainPage #search-box-container .search-box {
+//     display: flex;
+//     align-items: center;
+//     margin-top: 5px;
+//     font-family: "NanumGothic", serif;
+//     font-size: 10px;
+//     min-width: 800px;
+// }
+// .MainPage #button-container {
+//     margin: 0;
+//     text-align: left;
+//     width: auto;
+//     display: flex;
+//     align-items: start;
+//     justify-content: space-around;
+//     margin-top: 15px;
+//     min-width: 200px;
+// }
+// .MainPage #button-container .button {
+//     font-size: $button-size;
+//     // height: $button-size;
+//     width: $button-size;
+//     cursor: pointer;
+//     z-index: 100;
+// }
 .MainPage .Content {
     padding: 0 5vw;
     padding-bottom: 10px;
